Fix: Pacotão de correção de warnings e notices

+++ usuario/exportar_database.php
@@ -444,8 +444,9 @@ readfile($zip_name);
 unlink($zip_name);
 foreach($exportFiles as $file)
 {
-    unlink($file[0]);
-    
+    if (file_exists($file[0])) {
+        unlink($file[0]);
+    }
 }
 
 $usuario->atualizarDownload($userId);
